+++ example/main.dart
@@ -1,26 +1,20 @@
-# angular_sentry
+import 'dart:html' hide Event;
 
-Helper to implements sentry with Angular.
-
-## Usage
-
-### Basic
-
-```dart
 import "package:angular/angular.dart";
 import "package:angular_sentry/angular_sentry.dart";
+import 'package:http/http.dart';
 
 // ignore: uri_has_not_been_generated
 import 'main.template.dart' as ng;
 
 const sentryModule = Module(provide: [
-    ValueProvider.forToken(sentryLoggerToken, "MY_SENTRY_DSN"),
-    ValueProvider.forToken(sentryEnvironmentToken, "production"),
-    ValueProvider.forToken(sentryReleaseVersionToken, "1.0.0"),
-    ClassProvider<ExceptionHandler>(
-        ExceptionHandler,
-        useClass: AngularSentry,
-      ),
+  //ValueProvider.forToken(sentryLoggerToken, "MY_SENTRY_DSN"),
+  ValueProvider.forToken(sentryEnvironmentToken, "production"),
+  ValueProvider.forToken(sentryReleaseVersionToken, "1.0.0"),
+  ClassProvider<ExceptionHandler>(
+    ExceptionHandler,
+    useClass: AppSentry,
+  ),
 ]);
 
 @GenerateInjector(
@@ -29,26 +23,17 @@ const sentryModule = Module(provide: [
 const scannerApp = ng.scannerApp$Injector;
 
 main() {
-  runApp(appComponentNgFactory, createInjector: scannerApp);
+  runApp(ng.AppComponentNgFactory, createInjector: scannerApp);
 }
-```
 
-### Advanced
-
-Implement your own class using AngularSentry
-
-```dart
-
-const sentryModule = Module(provide: [
-    ...
-    ClassProvider<ExceptionHandler>(
-        ExceptionHandler,
-        useClass: AppSentry,
-      ),
-]);
-
-main() {
-  runApp(appComponentNgFactory, createInjector: scannerApp);
+@Component(
+  selector: 'app',
+  template: '<button (click)="error()">Catch error</button>',
+)
+class AppComponent {
+  void error() {
+    throw Exception("Test error");
+  }
 }
 
 class AppSentry extends AngularSentry {
@@ -56,7 +41,7 @@ class AppSentry extends AngularSentry {
       : super(
           injector,
           zone,
-          dsn: "MY_SENTRY_DSN",
+          //dsn: "MY_SENTRY_DSN",
           environment: "production",
           release: "1.0.0",
         );
@@ -78,5 +63,3 @@ class AppSentry extends AngularSentry {
     }
   }
 }
-
-```